CRUD Products + restricted pages
This way we can ,as an ADMIN, Add a product, Edit (update) it and delete it, products.php and dashboard.php Show datas for products

+++ Part_4/vitrine/template/admin/dashboard.php
@@ -1,44 +1,49 @@
 <?php
+// If constant is not defined error 403 + exit().
+if (!defined('ACCESS_GRANTED')) {
+    http_response_code(403);
+    exit();
+}
 
 $products = getAllProducts();
 
 if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_GET['action']) && $_GET['action'] == 'deletePostById' && isset($_POST['id'])) {
     
+
     $id = intval($_POST['id']);
     deleteProduct($id);
     header('location: index.php?route=admin/dashboard');
-}
+} 
 ?>
 <!-- Admin dashboard -->
 <h1>ADMIN DASHBOARD</h1>
 <h2>PRODUCTS</h2>
 <a href="index.php?route=admin/products" class="btn adminBtn">Add Product</a><br><br>
 
-
-<?php if(!isset($products[0])){?>
-<p>No product in Database</p>
-<?php } else{?>
-    <?php foreach($products as $product){?>
-    <div class="parent">
-        <!-- Form to show and edit Product -->
-        <form action="editProduct.php?route=editProduct&action=updateProductById&id=<?=$post['id']?>" method="POST">
-            <input type="hidden" id="id" name="id" value ="<?= $product['id'] ?>">
+<section class="productsAdmin">
+    <?php if(!isset($products[0])){?>
+    <p>No product in Database</p>
+    <?php } else{?>
+        <?php foreach($products as $product){?>
+        <div class="parent">
             <img src="assets/uploads/<?= $product['img'] ?>" alt="<?= $product['altImage']?>" style="height:150px; width:150px; border-radius:15px;">
-            <h3>Name : <?= $product['name']?></h3>
-            <p>Price : <?= $product['price']?> €</p>
-            <p>Description : <?= $product['description']?></p>
-            <div class="btn-field">
-                <input type="submit" class="btn btn-info" value="Edit this product">
+            <h3 class="productName">Name : <?= $product['name']?></h3>
+            <p class="productPrice">Price : <?= $product['price']?> €</p>
+            <p class="productDescription">Description : <?= $product['description']?></p>
+            
+            <!-- Form to delete Product -->
+            <div class="enfant">
+                <div class="btn-field">
+                    <a href="index.php?route=admin/update/editProduct&action=updateProductById&id=<?= $product['id']?>" class="btn">Edit</a>
+                </div>
+                <form method="POST" action="?route=admin/dashboard&action=deletePostById" onsubmit="return confirm('Voulez-vous confirmer la supression de <?= $product['name']?> ?');" style="display:inline;">
+                    <input type="hidden" name="id" value="<?= $product['id'] ?>">
+                    <div class="btn-field">
+                        <input type="submit" class="btn btn-red" value="Delete">
+                    </div>
+                </form>
             </div>
-        </form>
-        <!-- Form to delete Product -->
-        <form method="POST" action="?route=admin/dashboard&action=deletePostById" onsubmit="return confirm('Voulez-vous confirmer la supression de <?= $product['name']?> ?');" style="display:inline;">
-        
-            <input type="hidden" name="id" value="<?= $product['id'] ?>">
-            <div class="btn-field">
-                <input type="submit" class="btn btn-red" value="Delete this product">
-            </div>
-        </form>
-    </div>
+        </div>
+        <?php }?>
     <?php }?>
-<?php }?>
+</section>
